Make width and height variables short

+++ updating-terminal/loop.h
@@ -7,7 +7,7 @@
 
     #include <unistd.h> // used for usleep
 
-    void loop(int W, int H) {
+    void loop(short W, short H) {
         usleep(1000 * 1000/60); // 60 fps update
         printf("\033[%dA", H);  // snap cursor to first written line
         printf("\r");           // snap to start of line 
@@ -16,8 +16,8 @@
 #else
 
     #include <windows.h> // used for Sleep in milliseconds
-
-    void loop(int W, int H) {
+    
+    void loop(short W, short H) {
         Sleep(1000./60.); // 60 fps update
         HANDLE StdOut = GetStdHandle(STD_OUTPUT_HANDLE); // get stdout handle
         CONSOLE_SCREEN_BUFFER_INFO lpConsoleScreenBufferInfo;
